chore: refactor to target instend of modules
