+++ Features/Sources/PreferencesModule/Features.swift
@@ -1,0 +1,2 @@
+// The Swift Programming Language
+// https://docs.swift.org/swift-book
