--- PaywallModule/Package.swift
@@ -1,36 +1,0 @@
-// swift-tools-version: 6.1
-// The swift-tools-version declares the minimum version of Swift required to build this package.
-
-import PackageDescription
-
-let package = Package(
-    name: "PaywallModule",
-    platforms: [.iOS(.v16)],
-    products: [
-        // Products define the executables and libraries a package produces, making them visible to other packages.
-        .library(
-            name: "PaywallModule",
-            targets: ["PaywallModule"]
-        )
-    ],
-    dependencies: [
-        .package(path: "./DesignSystemModule"),
-        .package(path: "./CoreModule")
-    ],
-    targets: [
-        // Targets are the basic building blocks of a package, defining a module or a test suite.
-        // Targets can depend on other targets in this package and products from dependencies.
-        .target(
-            name: "PaywallModule",
-            dependencies: [
-                "DesignSystemModule",
-                "CoreModule"
-            ],
-            swiftSettings: [.swiftLanguageMode(.v5)]
-        ),
-        .testTarget(
-            name: "PaywallModuleTests",
-            dependencies: ["PaywallModule"]
-        )
-    ]
-)
